research/xidian/matchnet/requirements.txt
mindspore==1.8.1
moxing_framework==2.0.0.rc2.4b57a67b
PyYAML==6.0
scipy~=1.10.1
numpy >= 1.17.0
pillow~=9.4.0